+++ R/matrix.tools.R
@@ -1,3 +1,5 @@
+# matrix subset tools -----------------------------------------------------
+
 #' Non Diagonal Part of a Matrix
 #'
 #' Similarly to upper.tri and lower.tri, returns a matrix of logicals to identify the non-diagonal of a square matrix
@@ -58,6 +60,9 @@ diagonal<- function(M,output=c("matrix.logical","vector.values")) {
 #'
 non.zero.non.diag<- function(M) {which(M>0&!diagonal(M),arr.ind = TRUE,useNames = TRUE)}
 
+
+# matrix sum tool ---------------------------------------------------------
+
 #' Matrix sum removing NAs
 #' Equivalent to the elmement-wise matrix addition, but replacing NAs by zeros. Internal use
 #'
@@ -93,6 +98,8 @@ n.observed_edges<- function(scan_list,diag=0){
          )
   )
 }
+
+# Adjacency mode tools ----------------------------------------------------
 
 #' Make Adjacency fit the selected mode
 #' From a directed adjacency matrix, make it fit the selected mode.
@@ -160,5 +167,70 @@ binary_adjacency_mode<- function(Adj,mode = c("directed", "undirected", "max","m
          "upper" = ,
          "lower" =  Adj
   )
+}
+
+
+# obs.prob tools ----------------------------------------------------------
+
+#' "Reverse" matrix rows and column order
+#' For a matrix M with dim(M)=(n,m), swap element M[i,j] swapped by M[n-i+1,m-j+1]
+#'
+#' @param M a matrix
+#'
+#' @return a similarly dimensioned matrix with element M[i,j] swapped by M[n-i+1,m-j+1]
+#' @export
+#'
+#' @examples
+#' reverse_i.n(matrix(runif(15),3,5))
+reverse_i.n<- function(M){
+  Reversed<- matrix(0,nrow = nrow(M),ncol = ncol(M),dimnames = list(rownames(M),colnames(M)))
+  for(i in 1:nrow(M)){
+    for(j in 1:ncol(M)){
+      Reversed[i,j]<- M[nrow(M)-i+1,ncol(M)-j+1]
+    }
+  }
+  Reversed
+}
+
+#' Create probability of edge observation matrix
+#' Create probability of edge observation matrix with user-defined functions. Designed for trait-based or network-based biases.
+#'
+#' @param Adj a reference adjacency (square) matrix
+#' @param obs.prob_fun a function to be applied to calculate each cell of the output matrix. The function is applied to the row and column coordinates i and j (e.g. sum(i,j)), or is composed internally with bias_fun() like this: original_fun(bias_fun(i),bias_fun(j))
+#' @param bias_fun function to implement a bias into the edge observability probability. Can be "trait-based" if related to a node (and in the result to each of its edges), to an edge, or "network-based" if related to a network metric (e.g. a node's strength)
+#' @param reverse logical. Apply function reverse_i.n() to the output before returning it.
+#'
+#' @return a matrix of "probability" of edge obervability. The function can return a matrix of any real number (with zeros at the diagonal), and the scaling to [0,1] is currently handled in observable_edges().
+#' @export
+#'
+#' @examples
+#' set.seed(42)
+#' n<- 6;nodes<- as.character(1:n);
+#' total_scan<- 20;n.boot<- 5;
+#' focal.list<- sample(nodes,total_scan,replace = TRUE)
+#'
+#' Adj<- matrix(data = 0,nrow = n,ncol = n,dimnames = list(nodes,nodes))
+#' Adj[non.diagonal(Adj)]<- sample((0:round(total_scan*.50)),n*(n-1),replace = TRUE)
+#'
+#' traits<- rnorm(nrow(Adj),0,1)
+#' trait.bias_fun<- function(x) {traits[x]}
+#'
+#' obs.prob_bias(Adj,sum,bias_fun = trait.bias_fun)
+obs.prob_bias<- function(Adj,obs.prob_fun,bias_fun = NULL,reverse = FALSE){
+  n<- nrow(Adj);
+  obs.prob<- matrix(0,n,n,dimnames = list(rownames(Adj),colnames(Adj)));
+  if(!is.null(bias_fun)) {
+    original_fun<- obs.prob_fun;
+    obs.prob_fun<- function(i,j) {original_fun(bias_fun(i),bias_fun(j))};
+  }
+
+  for(i in seq_len(n)) {
+    for(j in seq_len(n)) {
+      if(i!=j){
+        obs.prob[i,j]<- obs.prob_fun(i,j);
+      }
+    }
+  }
+  if(!reverse) obs.prob else reverse_i.n(obs.prob)
 }